Binary Search - CN
****Link******
https://www.codingninjas.com/codestudio/problems/binary-search_972?leftPanelTab=0
*****Description*****
You have been given a sorted(in ascending order) integer array/list(ARR) of size N and an element X. Write a function to search this element in the given input array/list using 'Binary Search'. Return the index of the element in the input array/list. In case the element is not present in the array/list, then return -1.